+++ src/main/kotlin/com/nylas/models/EventQueryOrderBy.kt
@@ -1,0 +1,8 @@
+package com.nylas.models
+
+import com.squareup.moshi.Json
+
+enum class EventQueryOrderBy {
+  @Json(name = "start")
+  START,
+}