+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Nylas Java SDK Changelog
 
+### Unreleased
+* Fixed ListEventQueryParams to use a more specific type for orderBy parameter and correct JSON name.
+
 ### [2.8.0] - Release 2025-04-30
 * Added support for Notetaker APIs
 * Added support for Notetaker via the calendar and event APIs